prod 09/09/2025 à 10h17

+++ src/main/java/com/ged/projection/PerformancePortefeuilleActionnaireProjection.java
@@ -14,8 +14,8 @@ public interface PerformancePortefeuilleActionnaireProjection {
     LocalDateTime getDateOp();
     BigDecimal getSouscription();
     BigDecimal getRachat();
-    Long getPeriode();
-    Long getNbreJours();
+    BigDecimal getPeriode();
+    BigDecimal getNbreJours();
     BigDecimal getMontantPondere();
     BigDecimal getRevenuNet();
     BigDecimal getApportNet();
